Move useSource from src/utils/ to src/hooks/
