+++ src/hooks/useSubscription.ts
@@ -3,7 +3,8 @@ import { useCallback, useRef, useMemo } from 'react';
 import { pipe, concat, fromValue, switchMap, map, scan } from 'wonka';
 
 import { useClient } from '../context';
-import { useSource, useBehaviourSubject, CombinedError } from '../utils';
+import { CombinedError } from '../utils';
+import { useSource, useBehaviourSubject } from './useSource';
 import { OperationContext } from '../types';
 import { useRequest } from './useRequest';
 import { initialState } from './constants';
