+++ src/utils/index.ts
@@ -4,7 +4,6 @@ export * from './result';
 export * from './typenames';
 export * from './toSuspenseSource';
 export * from './stringifyVariables';
-export * from './useSource';
 export * from './withPromise';
 
 export const noop = () => {
